refactor: clean up AvroAttrsBridge implementation
To get it ready for prototype usage and for extentions
